Use sed to remove row numbers from delimiters
Prior to this, removing a row from somewhere within a table (i.e. not at
the end of the table) would result in dirty diff output because all the
row delimiters would appear to have changed.

The sed command matches delimiters in the format
`*************************** 2. row ***************************`, and
removes the row number from the line, making all of them identical.